feat Exception Handling

+++ urssu/src/main/kotlin/com/example/urssu/service/CommentService.kt
@@ -1,12 +1,14 @@
 package com.example.urssu.service
 
+import com.example.urssu.config.BaseException
+import com.example.urssu.config.BaseResponse
+import com.example.urssu.config.BaseResponseStatus
 import com.example.urssu.domain.entity.ArticleEntity
 import com.example.urssu.domain.entity.CommentEntity
 import com.example.urssu.domain.entity.UserEntity
 import com.example.urssu.domain.repository.ArticleRepository
 import com.example.urssu.domain.repository.CommentRepository
 import com.example.urssu.domain.repository.UserRepository
-import com.example.urssu.dto.ArticleReqDto
 import com.example.urssu.dto.CommentReqDto
 import com.example.urssu.dto.UserInfoDto
 import org.springframework.beans.factory.annotation.Autowired
@@ -24,31 +26,45 @@ class CommentService {
     @Autowired lateinit var commentRepository: CommentRepository
 
     fun postComment(commentReqDto: CommentReqDto, articleId: Int): CommentEntity {
-        val userEntity: UserEntity = userRepository.findByEmail(commentReqDto.email)
+        if(userRepository.findByEmailAndPassword(commentReqDto.email, commentReqDto.password).isEmpty) {
+            val baseException = BaseException(BaseResponseStatus.USER_EMPTY_USER)
+            throw baseException
+        }
+
+        if(articleRepository.findById(articleId).isEmpty){
+            val baseException = BaseException(BaseResponseStatus.POST_NOT_EXIST_POST)
+            throw baseException
+        }
+
+        val userEntity: UserEntity = userRepository.findByEmail(commentReqDto.email).get()
         val articleEntity: ArticleEntity = articleRepository.findById(articleId).get()
         return commentRepository.save(commentReqDto.toEntity(userEntity, articleEntity))
     }
 
     fun updateComment(commentReqDto: CommentReqDto, articleId: Int, commentId: Int): CommentEntity{
+        if(userRepository.findByEmailAndPassword(commentReqDto.email, commentReqDto.password).isEmpty) {
+            val baseException = BaseException(BaseResponseStatus.USER_EMPTY_USER)
+            throw baseException
+        }
+
         var commentEntity: CommentEntity =  commentRepository.findById(commentId).get()
-        commentEntity = updateCommentUsingDto(commentEntity, commentReqDto, articleId)
+        val userEntity: UserEntity = userRepository.findByEmailAndPassword(commentReqDto.email, commentReqDto.password).get()
+        val articleEntity: ArticleEntity = articleRepository.findById(articleId).get()
+
+        commentEntity.updateEntity(commentReqDto, userEntity, articleEntity)
         return commentRepository.save(commentEntity)
     }
 
     fun deleteComment(userInfoDto: UserInfoDto, articleId: Int, commentId: Int){
+        if(userRepository.findByEmailAndPassword(userInfoDto.email, userInfoDto.password).isEmpty) {
+            val baseException = BaseException(BaseResponseStatus.USER_EMPTY_USER)
+            throw baseException
+        }
+
         val comments: List<CommentEntity> = commentRepository.findAllByEmailAndPassword(userInfoDto.email, userInfoDto.password)
         for(comment in comments){
             if(comment.articleEntity.articleId == articleId && comment.commentId == commentId)
                 commentRepository.delete(comment)
         }
-    }
-
-    private fun updateCommentUsingDto(commentEntity: CommentEntity, commentReqDto: CommentReqDto, articleId: Int): CommentEntity{
-        val userEntity: UserEntity = userRepository.findByEmailAndPassword(commentReqDto.email, commentReqDto.password)
-        val articleEntity: ArticleEntity = articleRepository.findById(articleId).get()
-        commentEntity.content = commentReqDto.content
-        commentEntity.articleEntity = articleEntity
-        commentEntity.userEntity = userEntity
-        return commentEntity
     }
 }
